Loosen defensiveness
We still want to only enable monitoring for the first registrant, but it's not
necessary to restrict it to the first instance: we only need to restrict it to
one instance at a time.

In practice this won't make any difference to users, but it makes things easier
to test.